+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ---
 
-### 1.0
+# 1.0
 ```
-	* Replace WebSDK with Checkout SDK.
+    * Replace WebSDK with Checkout SDK.
     * Show the Order ID into the Success page URL.
     * Use Catalog currency for all Nuvei actions over the Order.
+    * Added plugin version checker.
 ```